Prefix any @-starting entries with ./
This is important for npm-packlist and others who may pass the file list
to a tar.create() command, since tar implementations treat @-prefixed
files as a tarball whose entries should be copied to the new file.

Currently npm/cli has some logic to do this, but it should really live
here, and not there.