fix: handling of nested unignores (#119)
The fix in #116 was not quite complete. git does allow unignoring
directories as long as its path to the root is unignored. This modifies
the check to handle this case. The test case in `nested-ignores.js` was
indeed wrong because it interpreted unignoring a specific file
`!/h/c/d/hcd` as unignoring the directory `!/h/c/d`. I added new tests
that specifically unignore a directory and therefore allow unignoring
anything else within it.

## References
Related to #116